refactor: split recipe list into subcomponents
